Split up Climate into Climate and ClimateMode
