+++ test/str_test.py
@@ -4,7 +4,7 @@ import unittest
 
 from xknx import XKNX
 from xknx.devices import (Action, ActionBase, ActionCallback, BinarySensor,
-                          Climate, Cover, DateTime, ExposeSensor, Light,
+                          Climate, ClimateMode, Cover, DateTime, ExposeSensor, Light,
                           Notification, RemoteValue, Scene, Sensor, Switch)
 from xknx.exceptions import (ConversionError, CouldNotParseAddress,
                              CouldNotParseKNXIP, CouldNotParseTelegram,
@@ -75,6 +75,20 @@ class TestStringRepresentations(unittest.TestCase):
             setpoint_shift_step=0.1,
             setpoint_shift_max=20,
             setpoint_shift_min=-20,
+            group_address_on_off='1/2/14',
+            group_address_on_off_state='1/2/15')
+        self.assertEqual(
+            str(climate),
+            '<Climate name="Wohnzimmer" temperature="None/GroupAddress("1/2/1")/None/None"  target_temperature="GroupAddress("1/2/2")/None/None/'
+            'None"  setpoint_shift="GroupAddress("1/2/3")/GroupAddress("1/2/4")/None/None" setpoint_shift_step="0.1" setpoint_shift_max="20" set'
+            'point_shift_min="-20" group_address_on_off="GroupAddress("1/2/14")/GroupAddress("1/2/15")/None/None" />')
+
+    def test_climate_mode(self):
+        """Test string representation of climate mode object."""
+        xknx = XKNX(loop=self.loop)
+        climate_mode = ClimateMode(
+            xknx,
+            name="Wohnzimmer Mode",
             group_address_operation_mode='1/2/5',
             group_address_operation_mode_state='1/2/6',
             group_address_operation_mode_protection='1/2/7',
@@ -83,17 +97,13 @@ class TestStringRepresentations(unittest.TestCase):
             group_address_controller_status='1/2/10',
             group_address_controller_status_state='1/2/11',
             group_address_controller_mode='1/2/12',
-            group_address_controller_mode_state='1/2/13',
-            group_address_on_off='1/2/14',
-            group_address_on_off_state='1/2/15')
+            group_address_controller_mode_state='1/2/13')
         self.assertEqual(
-            str(climate),
-            '<Climate name="Wohnzimmer" temperature="None/GroupAddress("1/2/1")/None/None"  target_temperature="GroupAddress("1/2/2")/None/None/'
-            'None"  setpoint_shift="GroupAddress("1/2/3")/GroupAddress("1/2/4")/None/None" setpoint_shift_step="0.1" setpoint_shift_max="20" set'
-            'point_shift_min="-20" group_address_operation_mode="GroupAddress("1/2/5")" group_address_operation_mode_state="GroupAddress("1/2/6")'
+            str(climate_mode),
+            '<ClimateMode name="Wohnzimmer Mode" '
+            'group_address_operation_mode="GroupAddress("1/2/5")" group_address_operation_mode_state="GroupAddress("1/2/6")'
             '" group_address_controller_status="GroupAddress("1/2/10")" group_address_controller_status_state="GroupAddress("1/2/11")" '
-            'group_address_controller_mode="GroupAddress("1/2/12")" group_address_controller_mode_state="GroupAddress("1/2/13")" '
-            'group_address_on_off="GroupAddress("1/2/14")/GroupAddress("1/2/15")/None/None" />')
+            'group_address_controller_mode="GroupAddress("1/2/12")" group_address_controller_mode_state="GroupAddress("1/2/13")" />')
 
     def test_cover(self):
         """Test string representation of cover object."""
